fix: update jQuery-related libraries, clean up latest versions
In many cases "latest" versions linked to Google CDN URLs that
were supposed to be continuously updated but that behavior has
stopped a while ago; as a result the links were often pointing
to older versions of libraries.

Git versions of jQuery are now hosted on releases.jquery.com,
there's also a separate 3.x-git version.